reorganized api folder structure
