+++ src/tapis-api/ml-hub/models/index.ts
@@ -1,6 +1,6 @@
 export { default as list } from './list';
 export { default as details } from './details';
-export { default as inferenceServerDetails } from './inferenceServerDetails';
+export { default as inferenceServerDetails } from './inference/inferenceServerDetails';
 export { default as listDownloadLinks } from './listDownloadLinks';
 export { default as listByAuthor } from './listByAuthor';
 export { default as listByDataset } from './listByDataset';
