+++ src/tapis-api/ml-hub/models/inference/inferenceServerDetails.ts
